+++ src/components/explore/exploreSlice.js
@@ -1,6 +1,6 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 
-import { instance } from "../utils";
+import { instance } from "../../utils";
 
 const initialState = {
   status: "idle",
