+++ src/components/post/postSlice.js
@@ -1,6 +1,5 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import axios from "axios";
-import { instance } from "../utils";
+import { instance } from "../../utils";
 
 const initialState = {
   status: "idle",
